Split React components into separate lib to implement per-version rendering.
