+++ .vscode/courier-web.code-workspace
@@ -8,6 +8,10 @@
       "path": "../@trycourier/courier-react"
     },
     {
+      "name": "courier-react-components",
+      "path": "../@trycourier/courier-react-components"
+    },
+    {
       "name": "courier-ui-inbox",
       "path": "../@trycourier/courier-ui-inbox"
     },
