+++ @trycourier/courier-react-components/README.md
@@ -1,0 +1,7 @@
+# `courier-react-components`
+
+> Adding Courier to a React app? Check out [`courier-react`](../courier-react/).
+
+This package contains React version agnostic components for Courier. `courier-react`
+uses the components here and provides a React version-specific render method. See
+
