+++ @trycourier/courier-react-components/src/context/render-context.ts
@@ -1,0 +1,5 @@
+import { createContext, ReactNode } from "react";
+
+type RenderFn = (node: ReactNode) => HTMLElement;
+
+export const CourierRenderContext = createContext<RenderFn | null>(null);
